+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   RTT echo support.
 - FIX: [hostmode] if tnc is already in hostmode, Flexpacket will not try so set it.
 - ADD: [experimental] filetransfere protocol YAPP, YAPPC, DIDADIT
+- ADD: [mheard] Show heard stations (view->Open MHeard)
 
 ## v0.8.1